refactored ui components with an additional blocks package, code cleanup
