+++ Code/src/main/java/nl/utwente/group10/ui/CustomUIPane.java
@@ -7,9 +7,9 @@ import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.scene.Node;
 import nl.utwente.ewi.caes.tactilefx.control.TactilePane;
-import nl.utwente.group10.ui.components.Block;
 import nl.utwente.group10.ui.components.Connection;
-import nl.utwente.group10.ui.components.DisplayBlock;
+import nl.utwente.group10.ui.components.blocks.Block;
+import nl.utwente.group10.ui.components.blocks.DisplayBlock;
 import nl.utwente.group10.ui.handlers.ConnectionCreationManager;
 
 /**
