+++ Code/src/main/java/nl/utwente/group10/ui/components/InputAnchor.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import nl.utwente.group10.haskell.expr.Expr;
 import nl.utwente.group10.haskell.expr.Ident;
 import nl.utwente.group10.ui.CustomUIPane;
+import nl.utwente.group10.ui.components.blocks.Block;
 import nl.utwente.group10.ui.handlers.AnchorHandler;
 
 /**
