+++ Code/src/main/java/nl/utwente/group10/ui/components/OutputAnchor.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.util.Optional;
 
 import nl.utwente.group10.ui.CustomUIPane;
+import nl.utwente.group10.ui.components.blocks.Block;
 import nl.utwente.group10.ui.handlers.AnchorHandler;
 
 /**
