+++ Code/src/main/java/nl/utwente/group10/ui/components/blocks/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains all of the Block components used by the UI
+ */
+
+package nl.utwente.group10.ui.components.blocks;
