+++ Code/src/test/java/nl/utwente/group10/ui/components/blocks/ComponentTest.java
@@ -1,4 +1,4 @@
-package nl.utwente.group10.ui.components;
+package nl.utwente.group10.ui.components.blocks;
 
 import javafx.application.Application;
 import javafx.stage.Stage;
